split rest-api part 2
