+++ src/components/dialogs/dynamicsimulation/dynamic-simulation-parameters-selector.js
@@ -24,7 +24,7 @@ import { useAutocompleteField } from '../../utils/inputs/use-autocomplete-field'
 import {
     fetchDynamicSimulationParameters,
     updateDynamicSimulationParameters,
-} from '../../../utils/rest-api';
+} from '../../../services/study/dynamic-simulation';
 
 export const checkDynamicSimulationParameters = (studyUuid) => {
     return fetchDynamicSimulationParameters(studyUuid).then((params) => {
