+++ src/components/dialogs/parameters/dynamicsimulation/dynamic-simulation-parameters.js
@@ -21,16 +21,16 @@ import {
     useParametersBackend,
     useStyles,
 } from '../parameters';
+import NetworkParameters from './network-parameters';
+import CurveParameters from './curve-parameters';
+import { fetchDynamicSimulationProviders } from '../../../../services/dynamic-simulation';
 import {
     fetchDefaultDynamicSimulationProvider,
     fetchDynamicSimulationParameters,
     fetchDynamicSimulationProvider,
     updateDynamicSimulationParameters,
     updateDynamicSimulationProvider,
-} from '../../../../utils/rest-api';
-import NetworkParameters from './network-parameters';
-import CurveParameters from './curve-parameters';
-import { fetchDynamicSimulationProviders } from '../../../../services/dynamic-simulation';
+} from '../../../../services/study/dynamic-simulation';
 
 const TAB_VALUES = {
     timeDelayParamsTabValue: 'TimeDelay',
