+++ src/components/dialogs/parameters/parameters.js
@@ -26,17 +26,6 @@ import {
 } from '@mui/material';
 
 import {
-    fetchDefaultSecurityAnalysisProvider,
-    fetchDefaultSensitivityAnalysisProvider,
-    fetchSecurityAnalysisProvider,
-    fetchSensitivityAnalysisProvider,
-    updateSecurityAnalysisProvider,
-    updateSensitivityAnalysisProvider,
-    getLoadFlowParameters,
-    getLoadFlowProvider,
-    getDefaultLoadFlowProvider,
-    setLoadFlowProvider,
-    setLoadFlowParameters,
     getSecurityAnalysisParameters,
     setSecurityAnalysisParameters,
 } from '../../../utils/rest-api';
@@ -70,6 +59,23 @@ import {
 } from '../../../services/loadflow';
 import { fetchSecurityAnalysisProviders } from '../../../services/security-analysis';
 import { fetchSensitivityAnalysisProviders } from '../../../services/sensitivity-analysis';
+import {
+    fetchDefaultSensitivityAnalysisProvider,
+    fetchSensitivityAnalysisProvider,
+    updateSensitivityAnalysisProvider,
+} from '../../../services/study/sensitivity-analysis';
+import {
+    getDefaultLoadFlowProvider,
+    getLoadFlowParameters,
+    getLoadFlowProvider,
+    setLoadFlowParameters,
+    setLoadFlowProvider,
+} from '../../../services/study/loadflow';
+import {
+    fetchDefaultSecurityAnalysisProvider,
+    fetchSecurityAnalysisProvider,
+    updateSecurityAnalysisProvider,
+} from '../../../services/study/security-analysis';
 
 export const CloseButton = ({ hideParameters, classeStyleName }) => {
     return (
