+++ src/components/results/dynamicsimulation/dynamic-simulation-result-tab.js
@@ -6,14 +6,14 @@
  */
 
 import { useNodeData } from '../../study-container';
-import {
-    fetchDynamicSimulationResult,
-    fetchDynamicSimulationResultTimeSeries,
-} from '../../../utils/rest-api';
 import WaitingLoader from '../../utils/waiting-loader';
 import DynamicSimulationResult from './dynamic-simulation-result';
 import { useCallback } from 'react';
 import { useSnackMessage } from '@gridsuite/commons-ui';
+import {
+    fetchDynamicSimulationResult,
+    fetchDynamicSimulationResultTimeSeries,
+} from '../../../services/study/dynamic-simulation';
 
 const dynamicSimulationResultInvalidations = ['dynamicSimulationResult'];
 const loadingMessage = 'LoadingRemoteData';
