+++ src/services/study/loadflow.js
@@ -1,0 +1,86 @@
+import {
+    backendFetch,
+    backendFetchJson,
+    backendFetchText,
+} from '../../utils/rest-api';
+import {
+    getStudyUrl,
+    getStudyUrlWithNodeUuid,
+    PREFIX_STUDY_QUERIES,
+} from './index';
+
+export function getDefaultLoadFlowProvider() {
+    console.info('get default load flow provier');
+    const getDefaultLoadFlowProviderUrl =
+        PREFIX_STUDY_QUERIES + '/v1/loadflow-default-provider';
+    console.debug(getDefaultLoadFlowProviderUrl);
+    return backendFetchText(getDefaultLoadFlowProviderUrl);
+}
+
+export function fetchLoadFlowInfos(studyUuid, currentNodeUuid) {
+    console.info(
+        `Fetching loadflow infos (status and result) for '${studyUuid}' and node '${currentNodeUuid}' ...`
+    );
+    const fetchLoadFlowInfosUrl =
+        getStudyUrlWithNodeUuid(studyUuid, currentNodeUuid) + '/loadflow/infos';
+    return backendFetchJson(fetchLoadFlowInfosUrl);
+}
+
+export function startLoadFlow(studyUuid, currentNodeUuid) {
+    console.info(
+        'Running loadflow on ' +
+            studyUuid +
+            ' and node ' +
+            currentNodeUuid +
+            '...'
+    );
+    const startLoadFlowUrl =
+        getStudyUrlWithNodeUuid(studyUuid, currentNodeUuid) + '/loadflow/run';
+    console.debug(startLoadFlowUrl);
+    return backendFetch(startLoadFlowUrl, { method: 'put' });
+}
+
+export function setLoadFlowParameters(studyUuid, newParams) {
+    console.info('set load flow parameters');
+    const setLoadFlowParametersUrl =
+        getStudyUrl(studyUuid) + '/loadflow/parameters';
+    console.debug(setLoadFlowParametersUrl);
+    return backendFetch(setLoadFlowParametersUrl, {
+        method: 'POST',
+        headers: {
+            Accept: 'application/json',
+            'Content-Type': 'application/json',
+        },
+        body: JSON.stringify(newParams),
+    });
+}
+
+export function getLoadFlowParameters(studyUuid) {
+    console.info('get load flow parameters');
+    const getLfParams = getStudyUrl(studyUuid) + '/loadflow/parameters';
+    console.debug(getLfParams);
+    return backendFetchJson(getLfParams);
+}
+
+export function getLoadFlowProvider(studyUuid) {
+    console.info('get load flow provider');
+    const getLoadFlowProviderUrl =
+        getStudyUrl(studyUuid) + '/loadflow/provider';
+    console.debug(getLoadFlowProviderUrl);
+    return backendFetchText(getLoadFlowProviderUrl);
+}
+
+export function setLoadFlowProvider(studyUuid, newProvider) {
+    console.info('set load flow provider');
+    const setLoadFlowProviderUrl =
+        getStudyUrl(studyUuid) + '/loadflow/provider';
+    console.debug(setLoadFlowProviderUrl);
+    return backendFetch(setLoadFlowProviderUrl, {
+        method: 'POST',
+        headers: {
+            Accept: 'application/json',
+            'Content-Type': 'application/json',
+        },
+        body: newProvider,
+    });
+}
