+++ src/services/study/security-analysis.js
@@ -1,0 +1,87 @@
+import {
+    backendFetch,
+    backendFetchJson,
+    backendFetchText,
+    getRequestParamFromList,
+} from '../../utils/rest-api';
+import {
+    getStudyUrl,
+    getStudyUrlWithNodeUuid,
+    PREFIX_STUDY_QUERIES,
+} from './index';
+
+export function startSecurityAnalysis(
+    studyUuid,
+    currentNodeUuid,
+    contingencyListNames
+) {
+    console.info(
+        `Running security analysis on ${studyUuid} and node ${currentNodeUuid} ...`
+    );
+
+    // Add params to Url
+    const contingencyListsQueryParams = getRequestParamFromList(
+        contingencyListNames,
+        'contingencyListName'
+    );
+    const urlSearchParams = new URLSearchParams(contingencyListsQueryParams);
+
+    const url = `${getStudyUrlWithNodeUuid(
+        studyUuid,
+        currentNodeUuid
+    )}/security-analysis/run?${urlSearchParams}`;
+
+    console.debug(url);
+    return backendFetch(url, { method: 'post' });
+}
+
+export function fetchSecurityAnalysisResult(studyUuid, currentNodeUuid) {
+    console.info(
+        `Fetching security analysis on ${studyUuid} and node ${currentNodeUuid} ...`
+    );
+    const url =
+        getStudyUrlWithNodeUuid(studyUuid, currentNodeUuid) +
+        '/security-analysis/result';
+    console.debug(url);
+    return backendFetchJson(url);
+}
+
+export function fetchSecurityAnalysisStatus(studyUuid, currentNodeUuid) {
+    console.info(
+        `Fetching security analysis status on ${studyUuid} and node ${currentNodeUuid} ...`
+    );
+
+    const url =
+        getStudyUrlWithNodeUuid(studyUuid, currentNodeUuid) +
+        '/security-analysis/status';
+    console.debug(url);
+    return backendFetchText(url);
+}
+
+export function fetchSecurityAnalysisProvider(studyUuid) {
+    console.info('fetch security analysis provider');
+    const url = getStudyUrl(studyUuid) + '/security-analysis/provider';
+    console.debug(url);
+    return backendFetchText(url);
+}
+
+export function updateSecurityAnalysisProvider(studyUuid, newProvider) {
+    console.info('update security analysis provider');
+    const url = getStudyUrl(studyUuid) + '/security-analysis/provider';
+    console.debug(url);
+    return backendFetch(url, {
+        method: 'POST',
+        headers: {
+            Accept: 'application/json',
+            'Content-Type': 'application/json',
+        },
+        body: newProvider,
+    });
+}
+
+export function fetchDefaultSecurityAnalysisProvider() {
+    console.info('fetch default security analysis provider');
+    const url = PREFIX_STUDY_QUERIES + '/v1/security-analysis-default-provider';
+    console.debug(url);
+    return backendFetchText(url);
+}
